Change career stats format to use nested objects instead of arrays
Now career stats can be accessed by accessing the career_stats object's
key with the hero name and then accessing that object with the stat's
name.

This is done by mapping each stat into name and value pairs and then
converting those pairs into an objects with keys of the stat name and
values of the stat value.