Consolidate getStatNameFromDropdown composition
Each place that getStatNameFromDropdown was used, it was composed with
cheerio so that you could pss cheerio arguments to the function instead
of a creating a cheerio object and then passing it.

Since this composition was being used multiple times, it's easier to do
the composition once and name the resulting function so that the
callsites can use that function.